Added prettier_help. Made more cogs
