+++ Botkeyy.py
@@ -2,12 +2,16 @@ import discord
 from discord.ext import commands
 import os
 from dotenv import load_dotenv
+from pretty_help import DefaultMenu, PrettyHelp
 
 load_dotenv()
 
 DISCORD_TOKEN = os.getenv('DISCORD_TOKEN')
 intents = discord.Intents().all()
 bot = commands.Bot(command_prefix="!", intents=intents, help_command=None)
+
+menu = DefaultMenu('◀️', '▶️', '❌')
+bot.help_command = PrettyHelp(navigation=menu, color=discord.Colour.green())
 
 @bot.event
 async def on_ready():
@@ -21,9 +25,7 @@ async def on_member_join(member):
         to_send = 'Welcome {0.mention} to {1.name}!'.format(member, guild)
         await guild.system_channel.send(to_send)
         
-extensions = ['cogs.HelpCommands',
-              'cogs.OnMessage',
-              'cogs.BotCommands']
+extensions = ['cogs.infoCommands','cogs.OnMessage','cogs.utilityCommands','cogs.BotCommands','cogs.funCommands']
 
 if __name__ == '__main__':
     for ext in extensions:
